+++ brewblox_ctl_lib/config_files/amd64/docker-compose.yml
@@ -8,6 +8,4 @@ services:
     labels:
       - "traefik.port=5000"
       - "traefik.frontend.rule=PathPrefix: /spark-one"
-    command: >
-      --name=spark-one 
-      --mdns-port=${BREWBLOX_PORT_MDNS}
+    command: --name=spark-one --mdns-port=${BREWBLOX_PORT_MDNS}
